// apps/api/src/common/enums/watering.enums.ts
export enum WateringShift {
  DAY = 'DAY',
  NIGHT = 'NIGHT',
}

export enum WateringType {
  AUTO = 'AUTO',
  WATER_CARRIER = 'WATER_CARRIER',
}

export enum WateringStatus {
  PLANNED = 'PLANNED',
  IN_PROGRESS = 'IN_PROGRESS',
  DONE = 'DONE',
  SKIPPED = 'SKIPPED',
  NEEDS_REVIEW = 'NEEDS_REVIEW',
}